refactor: 강제종료 버그 수정, 패키지 정리
