+++ app/src/main/java/woowacourse/shopping/model/CartProductMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.model
 
-import woowacourse.shopping.domain.CartProduct
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.CartProduct
+import woowacourse.shopping.domain.model.Product
 
 fun CartProduct.toUiModel(isChecked: Boolean = false, product: Product): CartProductModel =
     CartProductModel(isChecked, id, product.name, product.imageUrl, count, product.price.price)
