+++ app/src/main/java/woowacourse/shopping/model/ProductMapper.kt
@@ -1,5 +1,5 @@
 package woowacourse.shopping.model
 
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.Product
 
 fun Product.toUiModel(count: Int = 0) = ProductModel(id, name, imageUrl, price.price, count)
