+++ app/src/main/res/layout/item_cart_pagination.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="status"
-            type="woowacourse.shopping.domain.CartPageStatus" />
+            type="woowacourse.shopping.domain.cartsystem.CartPageStatus" />
 
         <variable
             name="onItemClick"
