+++ app/src/test/java/woowacourse/shopping/CartPresenterTest.kt
@@ -11,12 +11,12 @@ import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.shopping.domain.CartPageStatus
-import woowacourse.shopping.domain.CartProduct
-import woowacourse.shopping.domain.CartRepository
-import woowacourse.shopping.domain.Price
-import woowacourse.shopping.domain.Product
-import woowacourse.shopping.domain.ProductRepository
+import woowacourse.shopping.domain.cartsystem.CartPageStatus
+import woowacourse.shopping.domain.model.CartProduct
+import woowacourse.shopping.domain.repository.CartRepository
+import woowacourse.shopping.domain.model.Price
+import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.model.CartProductModel
 import woowacourse.shopping.view.cart.CartContract
 import woowacourse.shopping.view.cart.CartPresenter
