--- domain/src/main/java/woowacourse/shopping/domain/CartSystemProduct.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.domain
-
-data class CartSystemProduct(val cartProduct: CartProduct, val price: Price)
