+++ domain/src/main/java/woowacourse/shopping/domain/cartsystem/CartPageStatus.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.cartsystem
 
 data class CartPageStatus(
     val isPrevEnabled: Boolean,
