+++ domain/src/main/java/woowacourse/shopping/domain/cartsystem/CartSystemResult.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.cartsystem
 
 data class CartSystemResult(val totalPrice: Int, val totalCount: Int)
