+++ domain/src/main/java/woowacourse/shopping/domain/model/CartProduct.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 data class CartProduct(val id: Int, val count: Int)
