+++ domain/src/main/java/woowacourse/shopping/domain/model/Price.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 @JvmInline
 value class Price(val price: Int) {
