+++ domain/src/main/java/woowacourse/shopping/domain/model/Product.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 data class Product(val id: Int, val name: String, val imageUrl: String, val price: Price) {
     companion object
