+++ domain/src/main/java/woowacourse/shopping/domain/pagination/PrevPagination.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.pagination
 
 interface PrevPagination<T> {
     val isPrevEnabled: Boolean
