+++ domain/src/main/java/woowacourse/shopping/domain/repository/RecentViewedRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.repository
 
 
 interface RecentViewedRepository {
